put time conversion in own file
